refactor: update code and unit test cases (#3)
* refact: update unit test case

* update

* update

* update

* update

* update

* update

* update

* update

* update

* update

* update

* update

* update

* update

* update

* update

* update

* update

* update

---------

+++ core/src/main/java/ai/z/openapi/service/web_search/WebSearchDTO.java
@@ -14,7 +14,7 @@ public class WebSearchDTO {
 	 * Creation time
 	 */
 	@JsonProperty("created")
-	private Integer created;
+	private Long created;
 
 	/**
 	 * ID
